requirements.txt
# AI Trading Bot Requirements

# Core
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Trading
MetaTrader5>=5.0.36
ta>=0.10.0

# Utilities
pyyaml>=6.0
colorama>=0.4.5
joblib>=1.1.0

# Development
pytest>=7.0.0
